consolidate electron files into one dir
