+++ packages/electron/package.json
@@ -12,10 +12,7 @@
     "postinstall": "echo '@packages/electron needs: yarn build'",
     "lint": "eslint",
     "start": "./bin/cypress-electron",
-    "test": "yarn test-unit",
-    "test-debug": "yarn test-unit --inspect-brk=5566",
-    "test-unit": "mocha --reporter mocha-multi-reporters --reporter-options configFile=../../mocha-reporter-config.json",
-    "test-watch": "yarn test-unit --watch"
+    "test": "yarn vitest"
   },
   "dependencies": {
     "@packages/icons": "0.0.0-development",
@@ -29,9 +26,11 @@
   "devDependencies": {
     "@electron/fuses": "1.8.0",
     "@electron/packager": "18.3.6",
+    "@vitest/coverage-v8": "^3.2.4",
     "execa": "4.1.0",
     "mocha": "3.5.3",
-    "systeminformation": "^5.27.7"
+    "systeminformation": "^5.27.7",
+    "vitest": "^3.2.4"
   },
   "files": [
     "dist",
