+++ packages/electron/src/install.ts
