+++ packages/electron/src/paths.ts
