+++ packages/electron/src/print-node-version.ts
@@ -1,2 +1,3 @@
+/* eslint-disable no-console */
 console.log(process.version.replace('v', ''))
 process.exit(0)
