--- packages/electron/test/.eslintrc.json
@@ -1,5 +1,0 @@
-{
-  "extends": [
-    "plugin:@cypress/dev/tests"
-  ]
-}
