--- packages/electron/test/electron_spec.js
@@ -1,3 +1,0 @@
-describe('cypress-electron', () => {
-  it('is true', () => {})
-})
